refactor: 유저의 활동과 관련된 페이지 리팩토링
