+++ src/pages/MyPage/components/UserInfo/index.tsx
@@ -1,5 +1,5 @@
 import ProfileImage from 'components/common/ProfileImage';
-import PATH from 'constants/path';
+import getPath from 'utils/getPath';
 import IUserData from 'pages/MyPage/types';
 import * as S from './style';
 
@@ -18,15 +18,15 @@ function UserInfoSection({
       <S.Name>{nickname}</S.Name>
       <S.LoginInfo $platform={platform}>{email}</S.LoginInfo>
       <S.ActivitiesWrapper>
-        <S.Activity to={`.${PATH.registInfoPage}`}>
+        <S.Activity to={getPath('myPage', 'registedTrees')}>
           <p>{postedTreesCount}</p>
           <p>등록한 트리</p>
         </S.Activity>
-        <S.Activity to={`${PATH.saveTreePage}`}>
+        <S.Activity to={getPath('myPage', 'savedTrees')}>
           <p>{savedTreesCount}</p>
           <p> 저장한 트리</p>
         </S.Activity>
-        <S.Activity to={`${PATH.reviewPage}`}>
+        <S.Activity to={getPath('myPage', 'registedReviews')}>
           <p>{reviewsCount}</p>
           <p>내가 쓴 후기</p>
         </S.Activity>
