+++ src/pages/MyPage/hooks.ts
@@ -1,4 +1,5 @@
 import useApiQuery from 'hooks/useApiQuery';
+import { HTTPError } from 'error/HTTPError';
 import IUserData from './types';
 
 const useProfile = () => {
@@ -6,11 +7,11 @@ const useProfile = () => {
 
   // TODO: HTTPError로 바꿔주세요.
   if (isError) {
-    throw new Error('리뷰를 불러오는 데 오류가 발생했습니다.');
+    throw new HTTPError('리뷰를 불러오는 데 오류가 발생했습니다.');
   }
 
   if (!data && !isLoading) {
-    throw new Error('데이터가 없습니다!');
+    throw new HTTPError('불러온 사용자의 프로필 데이터가 존재하지 않습니다.');
   }
 
   return data;
