--- src/pages/RegistedReviewPage/hooks.ts
@@ -1,9 +1,0 @@
-import useApiQuery from 'hooks/useApiQuery';
-import { IReviewsAPIResponse } from './types';
-
-const useReviews = () => {
-  const { data } = useApiQuery<IReviewsAPIResponse>('v1/my/reviews');
-  return data;
-};
-
-export default useReviews;
